fix Dd.js & del SG.js, TG.js
